Middleware renamed `keep-alive`
