+++ lib/constants/constants.js
@@ -118,7 +118,7 @@ insertions,
   ROUTES: {
     ping: '/ping.json',
     healthcheck: '/healthcheck.json',
-    session: '/session.json'
+    keepAlive: '/keep-alive.json'
   },
   EDITABLE: getEnvVarAsBoolean(EDITABLE, false),
   // NB. if no ENV, should NUNJUCKSOPTIONS values be inverted automatically?
