+++ lib/constants/constants.unit.spec.js
@@ -32,7 +32,7 @@ test('When constants is called', t => {
     ROUTES: {
       ping: '/ping.json',
       healthcheck: '/healthcheck.json',
-      session: '/session.json'
+      keepAlive: '/keep-alive.json'
     },
     EDITABLE: false,
     NUNJUCKSOPTIONS: {
