+++ lib/middleware/middleware.js
@@ -3,7 +3,7 @@ const logger = require('./logger/logger')
 const auth = require('./auth/auth')
 const ping = require('./ping/ping')
 const healthcheck = require('./healthcheck/healthcheck')
-const session = require('./session/session')
+const keepAlive = require('./keep-alive/keep-alive')
 const referrer = require('./referrer/referrer')
 const csrf = require('./csrf/csrf')
 const robots = require('./robots/robots')
@@ -23,7 +23,7 @@ module.exports = {
   auth,
   ping,
   healthcheck,
-  session,
+  keepAlive,
   robots,
   referrer,
   csrf,
