--- lib/middleware/session/session.unit.spec.js
@@ -1,9 +1,0 @@
-const test = require('tape')
-
-const session = require('./session')
-
-test('When session is required ', t => {
-  t.equal(typeof session.init, 'function', 'it should export the init method')
-
-  t.end()
-})
