+++ lib/middleware/user-session/user-session.js
@@ -14,7 +14,7 @@ const {
   PLATFORM_ENV,
   ROUTES: {
     ping: PING_URL,
-    session: SESSION_URL,
+    keepAlive: KEEP_ALIVE_URL,
     healthcheck: HEALTHCHECK_URL
   }
 } = require('~/fb-runner-node/constants/constants')
@@ -25,25 +25,23 @@ let serviceSecret
 
 const fakeSession = {}
 
-const genid = (
-  userId = user.createUserId(),
-  userToken = user.createUserToken()
-) => {
+let sessionName
+let secure
+let cookiePath
+let cookieMaxAge
+
+function genid (userId = user.createUserId(), userToken = user.createUserToken()) {
   const userDigest = user.createUserDigest(userId, userToken, serviceSecret)
   const userDigestBundle = {
     userId,
     userToken,
     userDigest
   }
+
   return JSON.stringify(userDigestBundle)
 }
 
-let sessionName
-let secure
-let cookiePath
-let cookieMaxAge
-
-const setSessionCookie = (res, encryptedDigestBundle) => {
+function setSessionCookie (res, encryptedDigestBundle) {
   res.cookie(sessionName, encryptedDigestBundle, {
     httpOnly: true,
     secure,
@@ -52,28 +50,28 @@ const setSessionCookie = (res, encryptedDigestBundle) => {
   })
 }
 
-const createSessionCookie = (res, userId, userToken) => {
+function createSessionCookie (res, userId, userToken) {
   const digestBundle = genid(userId, userToken)
   const encryptedDigestBundle = aes256.encrypt(serviceSecret, digestBundle)
   setSessionCookie(res, encryptedDigestBundle)
   return encryptedDigestBundle
 }
 
-const createSessionAndRedirect = (res, redirect = 'page.start') => {
+function createSessionAndRedirect (res, redirect = 'page.start') {
   createSessionCookie(res)
   return res.redirect(getUrl(redirect))
 }
 
-const clearSessionCookie = (res) => {
+function clearSessionCookie (res) {
   res.clearCookie('sessionId')
 }
 
-const clearSessionAndRedirect = (res, redirect = 'page.start') => {
+function clearSessionAndRedirect (res, redirect = 'page.start') {
   clearSessionCookie(res)
   return res.redirect(getUrl(redirect))
 }
 
-const addUserIdTokenMethods = (req, userId, userToken) => {
+function addUserIdTokenMethods (req, userId, userToken) {
   req.getUserId = () => userId
   req.getUserToken = () => userToken
 }
@@ -96,7 +94,6 @@ function init (options = {}) {
   secure = sessionOptions.secure
   cookiePath = sessionOptions.cookiePath
   cookieMaxAge = sessionOptions.cookieMaxAge
-
   serviceSecret = sessionOptions.serviceSecret || '<NONE>'
 
   router.use(cookieParser())
@@ -105,17 +102,17 @@ function init (options = {}) {
     const { originalUrl } = req
 
     /*
-     *  Exclude requests to the ping, healthcheck, and session routes
+     *  Exclude requests to the ping, healthcheck, and keepAlive routes
      */
-    if (originalUrl !== PING_URL && originalUrl !== HEALTHCHECK_URL && originalUrl !== SESSION_URL) {
+    if (originalUrl !== PING_URL && originalUrl !== HEALTHCHECK_URL && originalUrl !== KEEP_ALIVE_URL) {
       let encryptedDigestBundle = req.cookies[sessionName]
 
       if (!encryptedDigestBundle) {
-      //  Brand new session
+        //  Brand new session
         req.newSession = true
         encryptedDigestBundle = createSessionCookie(res)
       } else {
-      // keep cookie rolling
+        // keep cookie rolling
         setSessionCookie(res, encryptedDigestBundle)
       }
