+++ lib/server/server-middleware.js
@@ -11,7 +11,7 @@ const {
   auth,
   ping,
   healthcheck,
-  session,
+  keepAlive,
   robots,
   referrer,
   csrf,
@@ -166,9 +166,7 @@ disallow: /public`
   }])
 
   // Create user session
-  middleware.push(userSession.init({
-    serviceSecret: SERVICE_SECRET
-  }))
+  middleware.push(userSession.init({ serviceSecret: SERVICE_SECRET }))
 
   // Ping route
   middleware.push(ping.init())
@@ -177,9 +175,7 @@ disallow: /public`
   middleware.push(healthcheck.init(options.validateHealthcheck, options))
 
   // Session route
-  middleware.push(session.init({
-    serviceSecret: SERVICE_SECRET
-  }))
+  middleware.push(keepAlive.init({ serviceSecret: SERVICE_SECRET }))
 
   // Add user data methods to req.user
   middleware.push(useAsync(userData.loadUserData))
